requirements.txt
requests==2.26.0
discord.py==2.4.0
Flask==3.1.0
asgiref==3.8.1